Add Python 3 support.
This approach creates two pairs of CPython and Python modules:
- `CPython2` and `Python2`
- `CPython3` and `Python3`

`Python2` and `Python3` compile the same source code,
`stdilb/public/Python/Python.swift`.

This separation of the Python modules is not ideal for other modules that want
to support both Python 2 and 3. Assuming pre-built packages will be released
with support for both Python 2 and 3, modules importing Python could be
compiled with either a `-DPYTHON2` or `-DPYTHON3` flag.

Example:

```swift
 // Default to Python 2.
 // Specify -DPYTHON3 to use Python 3.
 #if canImport(Python2) && PYTHON2
   import Python2
 #elseif canImport(Python3)
   import Python3
 #else
 #error("Toolchain does not support Python2 or Python3.")
 #endif

 // Use same `PythonObject` APIs.
```

TODO:
- Add Python 3 tests.