Lifted the restriction that when we partition and lowering a function foo(), and foo() references bar() as a function-typed attribute, bar() must have been partitioned and lowered already.
Added an e2e (runtime) test on dataset / iterator support.

To summarize the changes:

1. Changed TFPartition into be a module pass, to gain more flexibility in
iterating over the functions.

2. Implemented the following two-pass design for TFPartition, in order to lift
the ordering constraint on graph lowering.

The first pass partitions and lowers those partitionable functions into
graphs. When we process a function that references other functions that are not
yet lowered, such as foo() above, do not serialize the graph into bytes.

In the second pass, copy over the graph func bodies, such as copying the graph
function of bar() into the graph corresponding to foo() in the above example.

Note the following outstanding limitations in/after this PR:

1. Transitive referencies: If bar() references another graph function baz(),
baz() will need to be copied over to the graph of foo() as well, but currently
it might not, if baz() is processed after foo() in the first pass.

2. If bar() is defined in a different module from foo(), we will need to first
either load/link bar() when processing foo(), and lowering bar() (again) into a
graph function, or somehow load the graph function of bar() from its module, so
that we can copy it into the graph of foo().