Transform loops so that they have a single exit out of the loop header (#17865)
This is a first cut of a fix for SR-7765. This patch performs necessary transformations to convert loops into a form that can be translated directly to XF-TLA while loops. 

Note that this patch uses SILUndef values to deal with undefined values during the transformation. Therefore, the resulting CFG cannot yet be lowered to TF-XLA.