Sema: Simplify GatherConformancesListener
The inherited conformance check was dead because the type of a
conformance can never be an archetype.

Also, take care to substitute the conformance we are given, instead
of throwing it out and looking up a new one, because we want to be
careful to preserve any existing substitutions.

Part of the fix for <rdar://problem/40164371>.