[ObjectiveC] NSObject: Fix hashing implementation
NSObject defines its own interface for hashing and equality. Subclasses must implement custom hashing by overriding the hash property, or NSSet and NSDictionary will not work correctly.

To prevent a common source of errors, make the hashValue property Swift-only, and non-overridable. (It was declared as @objc open by historical accident.)

This is a potentially source-breaking change.