Attempt to fix non-deterministic failure in CursorInfo test
The problem appears to be that cursor info is using an unsound mechanism
(canUseASTWithSnapshots) to run without rebuilding the AST and this test
was triggering it unintentionally. By changing the order of the edits we
force cursor info to notice the tokens have changed only after the code
is semantically correct again. In real code this issue is rare and
solves itself the next time you invoke the cursor info.

rdar://42247603