Additional test updates for the ‘override’ and ‘dynamic’ requests.
Remove one SILGen test (!) because it is testing for something that
no longer happens: because Objective-C initializers now get properly
imported as being ‘dynamic’, we don’t have vtable entries for them
(or their overrides), so this test is no longer interesting. The
non-Objective-C version of this test still persists and is unchanged.