[AST] Try to resolve generic parameter depths lazily.
If we need the depth of a generic parameter declaration for
canonicalization, but it hasn’t been computed yet, resolve the
signature of its enclosing declaration. This is a heavy hammer for
a property that should be efficiently computable, but it’s along an
existing crashing path and addresses a compiler-crasher regression
due to us validating ahead of time with the request-evaluator.