Exclusivity: Disable verifier assert on noescape closure block args.
This is a small hole in the exclusivity model that verification is now catching
in 4.2. On 4.2, disable the verification just in this narrow, unsupported
case. This silences a harmless assert and reverts to the 4.2 behavior in this
unusual corner case. The exclusivity enforcement hole will be fully fixed on
master.

Fixed <rdar://problem/42242406> [SR-8266]: Compiler crash when checking
exclusivity of inout alias.