Sema: Check generic arguments applied to a type parameter
Previously we would just ignore generic arguments when applied to a
member type of a type parameter base. This was fine as long as we would
always do a final pass over a function's signature with the archetype
resolver, but when that goes away we have to check the generic arguments
here.

Most of the time, this is an error, unless the member type is a type
alias with generic parameters.