[Type checker] Move forcing of overridden/@objc/dynamic into finalizeType().
As part of factoring overridden decls/@objc/dynamic computation into
requests for the request-evaluator, I forced computation of this
information as part of requestMemberLayout. This is incorrect:
requestMemberLayout() is meant only to trigger finalization of the
enclosing class, which is responsible for making sure the type can be
properly laid out. Move the newly-introduced forcing logic into
finalizeType() where it belongs.

Thanks to @slavapestov for describing the intent here.