[ABI] Remove the “final” assertion from the ABI’s TrailingObjects.
In our ABI, we’re using some C++ ABI tricks to simplify the 
implementation of the runtime. One of them, in FullMetadata, requires
that we inherit from classes that we would also like to make use
TrailingObjects. Remove the “final” assertion from TrailingObjects
so we can do so.

We promise to be careful.