[update-checkout] Change dump-hashes to dump a update-checkout config file rather than a textual dump format.
Previously -dump-hashes dumped the current state of the hashes in a format like
the following:

$REPO_NAME $HASH $SUBJECT

While this dumps the relevant information, it is not machine parseable. Instead
in this commit, we dump out the hashes into a update-checkout config json file
with a scheme name of repro. Thus to reproduce the current repository state on
someone else's machine, you don't need to type in the exact hashes. Instead, one
can just do:

```
$ ./swift/utils/update-checkout -dump-hashes > repro.json
```

and then give the repro.json file to another swift developer. They then can run:

```
$ ./swift/utils/update-checkout -config=repro.json -scheme=repro
```

which will checkout all of the appropriate hashes for each repository to
reproduce the build on the other user's machine.