[ConstExpr] Emit traceback when failing during nested call evaluation (#18123)
When emitting a constexpr-evaluation failure message, this change emits
a trace of the calls the evaluator followed to generate the failure.

To make the output easier to read, calls on the same source line as the
failing location are elided from the output. So for example:

func foo() -> Bool {
  print("Un-constant")
  return true
}

test.swift:6:1: error: #assert condition not constant
^
test.swift:2:9: note: expression not evaluable as constant here
  print("Un-constant")

Notice there is no stack trace for the evaluation of the lambda.