[ParameterUpdate] Part 2: Parameterized protocol and synthesis.
The `Parameterized` protocol represents a type whose values have parameters.

The parameters of `Parameterized` types are stored properties marked with the
`@TFParameter` attribute.

For types that conform to `Parameterized`, the Swift compiler can synthesize a
member struct type `Parameters`, which includes all of the marked properties,
and a computed instance `allParameters`.

If all parameter types are the same, the compiler also synthesizes a
conformance of `Parameters` to `ParameterAggregate`.