[Runtime] Eliminate TargetProtocolConformanceRef::getName().
We don’t want the Swift runtime knowing the layout of Objective-C
protocols. Rather, it should use protocol_getName() when it needs the
(possibly mangled) name of an Objective-C protocol. Do so in the 
runtime. The code paths that handle ObjC vs. Swift protocols will
diverge significantly soon, and “getName()” isn’t an API we expect to
need in general.

Sink TargetObjCProtocolPrefix down into MetadataReader, because
that’s the only place where we want to expose the Objective-C protocol
layout, so we can form metadata without running code.