Added a new code path to create a TF graph for each module, instead of one for each partitionable function. The new code path is exercised by default, but can be turned off via a new flag "tf-module-level-graph". Turning it off is useful for some unit test files, where we check the GraphDef on a per test base basis, so we don't want all test cases in the file to end up in a single GraphDef for us to CHECK against.
Overall this PR simplifies the code base, as we no longer need to track the
graph function dependency, and copy over the dependent graph functions
later. Say `foo()` references `bar()` as a function-typed attribute. Previously,
when we lower foo(), if bar() has not been lowered, we need to track such
dependency, and come back later to copy the lowered graph function of `bar()`
into the graph of `foo()`. Now that `foo()` and `bar()` are lowered into the
same graph, the dependency tracking and graph function copying is no longer
needed.