When undef node creation fails, make sure an error diagnostic is emitted, so that optimizer pipeline / pass manager infra can later reject the program properly.
Otherwise, we are generating a binary where the serialized graph def has 0
bytes, and that triggers a runtime failure when running that graph.

With this patch, we get a compiler error diagnostic as expected:

```
$ ../build/Ninja-ReleaseAssert+stdlib-Release/swift-linux-x86_64/bin/swiftc  -Xllvm -tf-strict-deabstraction -Xllvm -tf-ensure-single-loop-exit=true -O  /usr/local/google/home/hongm/ssd_part/git/swift-models/MNIST/MNIST_s.swift
<unknown>:0: error: internal error generating TensorFlow graph:
Duplicate node name in graph: 'UndefConstDTensorHandleLFloatG'
```

Fixing the undef creation logic to avoid duplicate names (or even better,
removing the undef creation code) will be left for @bgogul to tackle.